# vp-bnd/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Videos API",
    "description": "API for managing videos",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
    "/videos": {
      "get": {
        "summary": "Get all videos",
        "responses": {
          "200": {
            "description": "A list of videos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new video",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "url"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The name of the video"
                  },
                  "description": {
                    "type": "string",
                    "description": "A brief description of the video"
                  },
                  "url": {
                    "type": "string",
                    "description": "The URL of the video"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Video created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{id}": {
      "get": {
        "summary": "Get a video by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single video",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          },
          "404": {
            "description": "Video not found"
          }
        }
      },
      "put": {
        "summary": "Update a video by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The name of the video"
                  },
                  "description": {
                    "type": "string",
                    "description": "A brief description of the video"
                  },
                  "url": {
                    "type": "string",
                    "description": "The URL of the video"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Video updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          },
          "404": {
            "description": "Video not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a video by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Video deleted"
          },
          "404": {
            "description": "Video not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Video": {
        "type": "object",
        "required": [
          "name",
          "description"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for a video"
          },
          "name": {
            "type": "string",
            "description": "The name of the video"
          },
          "description": {
            "type": "string",
            "description": "A brief description of the video"
          },
          "url": {
            "type": "string",
            "description": "The URL of the video"
          }
        }
      }
    }
  }
}